life_organizer/backend/requirements.txt
# Core Django
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# API Documentation
drf-spectacular==0.27.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File handling
Pillow==10.4.0

# Caching & Performance
redis==5.0.1
django-redis==5.4.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# AWS S3 (for production file storage)
boto3==1.34.131
django-storages==1.14.2

# Monitoring & Logging
sentry-sdk==1.40.6
django-prometheus==2.3.1

# Development & Testing
python-decouple==3.8
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==8.3.4
pytest-django==4.9.0
coverage==7.4.1

# Utilities
celery==5.3.4
flower==2.0.1
djangorestframework-simplejwt==5.3.0
pillow==10.4.0
django-filter==24.3
drf-spectacular==0.27.2
celery==5.3.4
redis==5.0.8
